Fix segfault when loaded without content.
My understanding is this is a combination of a frontend bug in
RetroArch and a core bug in lutro. RetroArch attempts to unload
content even when it was never loaded and lutro is deiniting content
in retro_deinit instead of retro_unload_core.

Please see:

https://github.com/libretro/RetroArch/issues/4493
https://github.com/libretro/RetroArch/pull/7102

Note that this fix in lutro is enough without also requiring the fix
in RetroArch which is needed for other cores.