tests/audio added: runs a series of state-machine based audio tests. (#195)
I didn't use the unit test framework for two reasons - one, I can't get it to work due to some lua include/require dir problems (could be windows-specific issue), and two, most of the "unit test framework" of luaunit is not useful for the purpose of testing the problems that usually need testing in multimedia applications.

I set this up as a SUCCESS/FAILURE thing at the moment for example, but often these things are better built as a fuzzy "just report # of voices and dump voice info" procedure, which is checked against log output that is verified as what we expect. And then, if a new log != expected log output, generate some error report for devs to look at. (and, by chance, many such "failures" are harmless or intentional, and verify new log as expected results and move on, etc).