make: assign config=player as default, fixes missing -O3 compiler optimization
Removed the use of `?=` assignment in gnu make, which doesn't work the way the documentation implies. The standard assignment of gnu make (`=`) is sufficient to the task of assigning a variable if and only if the variable was not previously assigned on the command line.

Added a check to throw an error if the config is missing