lutro: fix a whoooole bunch of prototype warnings (#264)
An interesting trivia of C99/C11 is that it isn't supposed to allow `()` for void functions. This is due to a classic rule of C language where `()` was actually a type of varadic, and matches any number of arguments. Thus, the use of either `(void)` or `(...)` is recommended for C programs to clearly differentiate them from the ambiguous `()` definition. (this doesn't apply to C++ code)