make: fix PhysFS defines to match expected assignment of [0,1]
When a codebase uses `#if HAVE_FOOBAR` it is necessary to use `HAVE_FOOBAR=1` to ensure the compiler actually enables the codeblock in the expected way. When the code uses `#ifdef HAVE_FOOBAR` then a define without any assignment is sufficient to enable the feature.

In general the `#if` form should be preferred over the `#ifdef` form for reasons of explicitness, along with a few other minor benefits related to inheriting or overriding the set value later on in the build system or at the source code level. PhysFs uses/expects this form, but lutro generally does not (yet).