make: fix PhysFS defines to match expected assignment of [0,1] (#270)
* make: fix PhysFS defines to match expected assignment of [0,1]

When a codebase uses `#if HAVE_FOOBAR` it is necessary to use `HAVE_FOOBAR=1` to ensure the compiler actually enables the codeblock in the expected way. When the code uses `#ifdef HAVE_FOOBAR` then a define without any assignment is sufficient to enable the feature. PhysFS uses `#if` and not `#ifdef`

* make: allow WARN_FLAGS to be read from the environment as well as from the CLI

... the power of `?=` unleashed.